chore: add jsx-nbsp autofix workflow and eslint rule
