+++ package.json
@@ -36,6 +36,7 @@
     "@vitejs/plugin-react": "^6.0.1",
     "eslint": "^9",
     "eslint-config-next": "^16.2.6",
+    "eslint-plugin-jsx-nbsp": "github:digitalby-me/jsx-nbsp-lint#v1",
     "jsdom": "^29.0.1",
     "tailwindcss": "^4",
     "typescript": "^5",
